part1/workout10/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
python-jose[cryptography]==3.3.0
